refactor: move filter components to molecules package in a new file Filter
- Also move respective tests to FilterTest
